+++ tests/test_skoolhtml.py
@@ -41,6 +41,7 @@ REF_SECTIONS = {
 MINIMAL_REF_FILE = """
 [Game]
 AddressAnchor={{address}}
+AsmSinglePage=0
 Created=
 LinkInternalOperands=0
 LinkOperands=CALL,DEFW,DJNZ,JP,JR
@@ -53,6 +54,7 @@ UDGFilename=udg{{addr}}_{{attr}}x{{scale}}
 METHOD_MINIMAL_REF_FILE = """
 [Game]
 AddressAnchor={{address}}
+AsmSinglePage=0
 Created=
 LinkInternalOperands=0
 LinkOperands=CALL,DEFW,DJNZ,JP,JR
@@ -69,6 +71,7 @@ CodeFiles={{address}}.html
 SKOOL_MACRO_MINIMAL_REF_FILE = """
 [Game]
 AddressAnchor={{address}}
+AsmSinglePage=0
 Created=
 LinkInternalOperands=0
 LinkOperands=CALL,DEFW,DJNZ,JP,JR
@@ -1956,7 +1959,7 @@ class SkoolMacroTest(HtmlWriterTestCase, CommonSkoolMacroTest):
         self._assert_error(writer, '#R$ABCD@main', 'Could not find instruction at $ABCD', prefix)
 
     def test_macro_r_asm_single_page(self):
-        ref = '[Game]\nAsmSinglePageTemplate=asm_single_page'
+        ref = '[Game]\nAsmSinglePage=1'
         skool = 'c40000 LD A,B\n 40001 RET'
         writer = self._get_writer(ref=ref, skool=skool)
 
@@ -2026,7 +2029,7 @@ class SkoolMacroTest(HtmlWriterTestCase, CommonSkoolMacroTest):
     def test_macro_r_other_code_asm_single_page(self):
         ref = """
             [Game]
-            AsmSinglePageTemplate=asm_single_page
+            AsmSinglePage=1
             [OtherCode:other]
         """
         writer = self._get_writer(ref=ref, skool='')
@@ -4730,7 +4733,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
         self._assert_files_equal(join(ASMDIR, '30000.html'), subs)
 
     def test_write_asm_entries_on_single_page(self):
-        ref = '[Game]\nAsmSinglePageTemplate=asm_single_page'
+        ref = '[Game]\nAsmSinglePage=1'
         skool = """
             ; Routine at 32768
             c32768 CALL 32775
@@ -4809,7 +4812,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
         title = 'The disassembly'
         ref = """
             [Game]
-            AsmSinglePageTemplate=asm_single_page
+            AsmSinglePage=1
             [Paths]
             AsmSinglePage={}
             [Titles]
@@ -5615,7 +5618,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             writer.write_map('MemoryMap')
 
     def test_write_map_with_single_page_template(self):
-        ref = '[Game]\nAsmSinglePageTemplate=asm_single_page'
+        ref = '[Game]\nAsmSinglePage=1'
         skool = """
             ; A routine here
             c32768 RET
@@ -5664,7 +5667,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
         ref = """
             [Game]
             AddressAnchor={address:04X}
-            AsmSinglePageTemplate=asm_single_page
+            AsmSinglePage=1
             [MemoryMap:MemoryMap]
             Intro=First instruction at #R43981.
         """
@@ -5705,7 +5708,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
         path = 'disassembly.html'
         ref = """
             [Game]
-            AsmSinglePageTemplate=asm_single_page
+            AsmSinglePage=1
             [Paths]
             AsmSinglePage={}
         """.format(path)
@@ -6265,7 +6268,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
         code_id = 'other'
         ref = """
             [Game]
-            AsmSinglePageTemplate=asm_single_page
+            AsmSinglePage=1
             [OtherCode:{}]
         """.format(code_id)
         other_skool = """
@@ -6333,7 +6336,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
         header = 'The disassembly'
         ref = """
             [Game]
-            AsmSinglePageTemplate=asm_single_page
+            AsmSinglePage=1
             [Paths]
             {0}-AsmSinglePage={1}
             [Titles]
@@ -6418,7 +6421,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
         code_id = 'other'
         ref = """
             [Game]
-            AsmSinglePageTemplate=asm_single_page
+            AsmSinglePage=1
             [OtherCode:{}]
         """.format(code_id)
         routine_title = 'Other code'
@@ -6460,7 +6463,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
         asm_single_page = 'disassembly.html'
         ref = """
             [Game]
-            AsmSinglePageTemplate=asm_single_page
+            AsmSinglePage=1
             [Paths]
             {0}-AsmSinglePage={0}/{1}
             [OtherCode:{0}]
@@ -8526,7 +8529,7 @@ class HtmlTemplateTest(HtmlWriterOutputTestCase):
         """
         ref = """
             [Game]
-            AsmSinglePageTemplate=asm_single_page
+            AsmSinglePage=1
 
             [Template:AsmSinglePage]
             <# foreach($entry,entries) #>
@@ -8560,7 +8563,7 @@ class HtmlTemplateTest(HtmlWriterOutputTestCase):
             [OtherCode:{}]
 
             [Game]
-            AsmSinglePageTemplate=asm_single_page
+            AsmSinglePage=1
 
             [Template:{}-AsmSinglePage]
             <# foreach($entry,entries) #>