+++ sphinx/source/changelog.rst
@@ -17,6 +17,8 @@ Changelog
   a disassembly page (as defined by the ``Asm-*`` parameters in the
   :ref:`titles` and :ref:`pageHeaders` sections)
 * Added the ``GameIndex`` parameter to the :ref:`pageHeaders` section
+* Replaced the ``AsmSinglePageTemplate`` parameter with the ``AsmSinglePage``
+  parameter in the :ref:`ref-Game` section
 
 7.2 (2019-06-02)
 ----------------